// attached_assets/extracted/grupo-tmseg/types.ts
export interface ProviderData {
  id: string;
  name: string;
  trading_name?: string; 
  cnpj: string;
  type: 'Escolta Caracterizada' | 'Pronta Resposta' | 'Moto Velada';
  contactName: string;
  status: 'Ativo' | 'Bloqueado' | 'Alvará Vencido';
  vehicleCount?: number;
  agentCount?: number;
  address?: string;
  city?: string;
  state?: string;
  alvaraValidity?: string; 
  alvaraUrl?: string;
}